Add like action for post index
