+++ src/actions/Like.js
@@ -1,0 +1,8 @@
+import * as types from 'constants/actionTypes/PostsActionTypes';
+
+const addLike = (id) => ({
+  type: types.ADD_LIKE_SUCCESS,
+  id
+});
+
+export default addLike;
