+++ src/constants/actionTypes/PostsActionTypes.js
@@ -1,3 +1,4 @@
 export const FETCH_POSTS_REQUEST = 'FETCH_POSTS_REQUEST';
 export const FETCH_POSTS_SUCCESS = 'FETCH_POSTS_SUCCESS';
 export const FETCH_POSTS_ERROR = 'FETCH_POSTS_ERROR';
+export const ADD_LIKE_SUCCESS = 'ADD_LIKE_SUCCESS';
