+++ src/containers/PostsContainer.js
@@ -1,5 +1,6 @@
 import BlogList from 'components/widgets/blog/List';
 import { connect } from 'react-redux';
+import addLike from 'actions/Like';
 
 const stateToProps = (state) => ({
   posts: state.posts.entries,
@@ -7,4 +8,10 @@ const stateToProps = (state) => ({
   error: state.posts.error
 });
 
-export default connect(stateToProps)(BlogList);
+const mapDispatchToProps = (dispatch) => ({
+  addLike: (id) => {
+    dispatch(addLike(id));
+  }
+});
+
+export default connect(stateToProps, mapDispatchToProps)(BlogList);
