+++ src/reducers/Post.js
@@ -10,7 +10,7 @@ const initialState =  {
 export default function(state = initialState, action) {
   switch (action.type) {
     case types.FETCH_POST_REQUEST:
-      return assign({}, initialState, { isFetching: true});
+      return assign({}, initialState, { isFetching: true });
     case types.FETCH_POST_ERROR:
       return assign({}, initialState, { error: true });
     case types.FETCH_POST_SUCCESS:
